Pagination: Align code-style and filenames
